requirements.txt
asgiref==3.7.2
attrs==23.2.0
bootstrap4==0.1.0
boto3==1.34.13
botocore==1.34.13
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.9.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.7
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-import-export==3.3.5
django-paginator==0.2a3
django-phonenumber-field==7.3.0
django-rest-knox==4.2.0
django-storages==1.14.2
djangorestframework==3.15.1
emails==0.6
et-xmlfile==1.1.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
import-export==0.3.1
install==1.3.5
jmespath==1.0.1
lxml==5.1.0
MarkupPy==1.14
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
parameters==0.2.1
parsel==1.8.1
phonenumbers==8.13.27
Pillow==10.1.0
premailer==3.10.0
psycopg2==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
resend==0.7.2
s3transfer==0.10.0
selenium==4.16.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stripe==8.3.0
tablib==3.5.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
users==1.0.dev0
w3lib==2.1.2
whitenoise==6.6.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
